--- api/controller/address.go
@@ -1,52 +1,0 @@
-package controller
-
-import (
-	"context"
-
-	"github.com/filecoin-project/go-address"
-	shared "github.com/filecoin-project/venus/venus-shared/types"
-	types "github.com/filecoin-project/venus/venus-shared/types/messager"
-
-	"github.com/filecoin-project/venus-messager/service"
-)
-
-type Address struct {
-	BaseController
-	AddressService *service.AddressService
-}
-
-func (a Address) SaveAddress(ctx context.Context, address *types.Address) (shared.UUID, error) {
-	return a.AddressService.SaveAddress(ctx, address)
-}
-
-func (a Address) GetAddress(ctx context.Context, addr address.Address) (*types.Address, error) {
-	return a.AddressService.GetAddress(ctx, addr)
-}
-
-func (a Address) HasAddress(ctx context.Context, addr address.Address) (bool, error) {
-	return a.AddressService.HasAddress(ctx, addr)
-}
-
-func (a *Address) ListAddress(ctx context.Context) ([]*types.Address, error) {
-	return a.AddressService.ListAddress(ctx)
-}
-
-func (a Address) UpdateNonce(ctx context.Context, addr address.Address, nonce uint64) error {
-	return a.AddressService.UpdateNonce(ctx, addr, nonce)
-}
-
-func (a Address) DeleteAddress(ctx context.Context, addr address.Address) error {
-	return a.AddressService.DeleteAddress(ctx, addr)
-}
-
-func (a Address) ForbiddenAddress(ctx context.Context, addr address.Address) error {
-	return a.AddressService.ForbiddenAddress(ctx, addr)
-}
-
-func (a Address) ActiveAddress(ctx context.Context, addr address.Address) error {
-	return a.AddressService.ActiveAddress(ctx, addr)
-}
-
-func (a Address) SetSelectMsgNum(ctx context.Context, addr address.Address, num uint64) error {
-	return a.AddressService.SetSelectMsgNum(ctx, addr, num)
-}
