--- api/controller/controller.go
@@ -1,7 +1,0 @@
-package controller
-
-import "github.com/filecoin-project/venus-messager/log"
-
-type BaseController struct {
-	Logger *log.Logger
-}
