+++ cli/cmd.go
@@ -1,23 +1,21 @@
 package cli
 
 import (
-	"context"
-	"github.com/filecoin-project/go-address"
 	"github.com/filecoin-project/go-jsonrpc"
 	"github.com/filecoin-project/venus-messager/service"
-	"github.com/filecoin-project/venus-messager/utils/actor_parser"
-	"github.com/filecoin-project/venus/venus-shared/types"
+	v1 "github.com/filecoin-project/venus/venus-shared/api/chain/v1"
 	"github.com/ipfs-force-community/venus-common-utils/apiinfo"
 	"github.com/urfave/cli/v2"
 
-	"github.com/filecoin-project/venus-messager/api/client"
 	"github.com/filecoin-project/venus-messager/config"
+
+	"github.com/filecoin-project/venus/venus-shared/api/messager"
 )
 
-func getAPI(ctx *cli.Context) (client.IMessager, jsonrpc.ClientCloser, error) {
+func getAPI(ctx *cli.Context) (messager.IMessager, jsonrpc.ClientCloser, error) {
 	cfg, err := config.ReadConfig(ctx.String("config"))
 	if err != nil {
-		return &client.Message{}, func() {}, err
+		return nil, func() {}, err
 	}
 
 	apiInfo := apiinfo.NewAPIInfo(cfg.API.Address, cfg.JWT.Local.Token)
@@ -26,35 +24,15 @@ func getAPI(ctx *cli.Context) (client.IMessager, jsonrpc.ClientCloser, error) {
 		return nil, nil, err
 	}
 
-	client, closer, err := client.NewMessageRPC(ctx.Context, addr, apiInfo.AuthHeader())
+	client, closer, err := messager.NewIMessagerRPC(ctx.Context, addr, apiInfo.AuthHeader())
 
 	return client, closer, err
 }
 
-func getNodeAPI(ctx *cli.Context) (*service.NodeClient, jsonrpc.ClientCloser, error) {
+func getNodeAPI(ctx *cli.Context) (v1.FullNode, jsonrpc.ClientCloser, error) {
 	cfg, err := config.ReadConfig(ctx.String("config"))
 	if err != nil {
-		return &service.NodeClient{}, func() {}, err
+		return nil, func() {}, err
 	}
 	return service.NewNodeClient(ctx.Context, &cfg.Node)
-}
-
-func getActorGetter(ctx *cli.Context) (actor_parser.ActorGetter, jsonrpc.ClientCloser, error) {
-	nodeAPI, closer, err := getNodeAPI(ctx)
-	if err != nil {
-		return nil, nil, err
-	}
-	return &getter{NodeClient: nodeAPI}, closer, nil
-}
-
-type getter struct {
-	*service.NodeClient
-}
-
-func (g *getter) StateGetActor(ctx context.Context, addr address.Address, tsKey types.TipSetKey) (*types.Actor, error) {
-	return g.NodeClient.StateGetActor(ctx, addr, tsKey)
-}
-
-func (g *getter) StateLookupID(ctx context.Context, addr address.Address, tsKey types.TipSetKey) (address.Address, error) {
-	return g.NodeClient.StateLookupID(ctx, addr, tsKey)
 }
